backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
Pillow
imagehash
networkx
numpy
scipy
scikit-learn
pydantic
aiofiles
httpx
python-dotenv
